pyproject.toml
[project]
name = "ai-service-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "keras>=3.7.0",
    "pillow>=11.0.0",
    "waitress>=3.0.2",
    "flask-sqlalchemy-lite>=0.1.0",
    "jax>=0.4.38",
    "onnx>=1.17.0",
    "onnxruntime>=1.20.1",
]

[dependency-groups]
dev = [
    "coverage>=7.6.8",
    "pytest>=8.3.4",
]

[build-system]
requires = ['hatchling']
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ['ai_service_platform']

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

# [tool.pyright]
# typeCheckingMode = "off"

[tool.ruff]
line-length = 88
